[ENG-2997]: Add types to EditorToast Actions
