+++ app/src/store/slices/global/editorToast/types.ts
@@ -1,0 +1,11 @@
+export interface EditorToastState {
+  editorToast: {
+    [key: string]: {
+      id?: string;
+      message?: string;
+      type?: string;
+      autoClose?: number;
+      scriptId?: string;
+    };
+  };
+}
