+++ app/src/store/slices/global/types.ts
@@ -1,6 +1,8 @@
+import { EditorToastState } from "./editorToast/types";
 import { GlobalModals } from "./modals/types";
 
 export interface GlobalSliceState {
   activeModals: GlobalModals;
+  editorToast: EditorToastState["editorToast"];
   [key: string]: any;
 }
